refectored query check into the util so it is called at same time as the main fetch
